+++ sourceCode/collections/collection.h
@@ -29,7 +29,7 @@ class CCollection : public Obj
     void serialize(CSer& ar);
     void performCollectionLoadingMapping(const std::map<int, int>* map, int opType);
     void performObjectLoadingMapping(const std::map<int, int>* map, int opType);
-    bool announceObjectWillBeErased(long long int objectHandle, bool copyBuffer);
+    bool announceObjectWillBeErased(int64_t objectHandle, bool copyBuffer);
     bool announceScriptStateWillBeErased(int detachedScriptHandle, bool simulationScript, bool sceneSwitchPersistentScript);
     CCollection* copyYourself() const;
     void emptyCollection();
@@ -41,7 +41,7 @@ class CCollection : public Obj
     std::string getCollectionName() const;
     bool getOverridesObjectMainProperties() const;
 
-    bool setCollectionHandle(long long int newHandle);
+    bool setCollectionHandle(int64_t newHandle);
     bool setCollectionName(const char* newName, bool check);
     bool setOverridesObjectMainProperties(bool o);
 
@@ -53,10 +53,10 @@ class CCollection : public Obj
     void pushCreationEvent();
 
     int getBoolProperty(const char* pName, bool& pState) const override;
-    int getLongProperty(const char* pName, long long int& pState) const override;
-    int getHandleProperty(const char* pName, long long int& pState) const override;
+    int getLongProperty(const char* pName, int64_t& pState) const override;
+    int getHandleProperty(const char* pName, int64_t& pState) const override;
     int getStringProperty(const char* pName, std::string& pState) const override;
-    int getHandleArrayProperty(const char* pName, std::vector<long long int>& pState) const override;
+    int getHandleArrayProperty(const char* pName, std::vector<int64_t>& pState) const override;
     int getStringArrayProperty(const char* pName, std::vector<std::string>& pState) const override;
     int getPropertyName(int& index, std::string& pName, std::string& appartenance, int excludeFlags) const override;
     int getPropertyInfo(const char* pName, int& info, std::string& infoTxt) const override;
