+++ sourceCode/interfaces/interfaceStack/interfaceStackHandleArray.h
@@ -5,7 +5,7 @@
 class CInterfaceStackHandleArray : public CInterfaceStackObject
 {
   public:
-    CInterfaceStackHandleArray(const long long int* vals, size_t cnt);
+    CInterfaceStackHandleArray(const int64_t* vals, size_t cnt);
     virtual ~CInterfaceStackHandleArray();
 
     CInterfaceStackObject* copyYourself() const override;
@@ -15,10 +15,10 @@ class CInterfaceStackHandleArray : public CInterfaceStackObject
     unsigned int createFromData(const char* data, unsigned char version, std::vector<CInterfaceStackObject*>& allCreatedObjects) override;
     static bool checkCreateFromData(const char* data, unsigned int& w, unsigned int l, unsigned char version);
 
-    const long long int* getValue(size_t* cnt) const;
-    void setValue(const long long int* vals, size_t cnt);
+    const int64_t* getValue(size_t* cnt) const;
+    void setValue(const int64_t* vals, size_t cnt);
     void setValue(const int* vals, size_t cnt);
 
   protected:
-    std::vector<long long int> _handles;
+    std::vector<int64_t> _handles;
 };
